AppProdutosMedia

+++ exercicios01/src/application/AppProdutosMedia.java
@@ -26,7 +26,7 @@ public class AppProdutosMedia {
 			System.out.println();
 		}
 		
-		int sum = 0;
+		double sum = 0.0;
 		
 		for(int i = 0; i < vect.length; i++) {
 			sum+= vect[i].getPrice();
